requirements.txt
Django>=4.2
gunicorn
psycopg2-binary
Pillow
djangorestframework
